+++ CHANGELOG.md
@@ -9,3 +9,4 @@
   * Displayname rendering was moved to the
     [ep\_user\_displayname](https://github.com/ether/ep_user_displayname#readme)
     plugin and the `permit_displayname_change` setting was removed.
+  * The plugin is now much more strict about what it accepts as valid settings.